Improve commit range validation
Description
===========

It is possible to provide a valid tag name or commit sha as _from_ and
_to_ parameters. The resulting commits are checked if they are available
but not if they are reachable from the provided branch.

```
NR
9         * commit 4 on develop
8         * commit 3 on develop
7       * | commit 5 (tag: v0.1.1)
6       * | commit 4
5       | * commit 2 on develop
4       * | commit 3 (tag: v0.1.0)
3       | * commit 1 on develop
        |/
2       * commit 2
1       * Initial commit
```

One could call `changeDetector.detectChangesFromTag('v0.1.0`, 'v0.1.1', 'develop')`
or with the commit SHA's and the function would fail with an generic
`java.lang.IllegalArgumentException`. This patch adds a range check for
the provided parameters and throws a more meaningfill exception.

Changes
=======

* ![IMPROVE] range validation for provided from/to parameters