Add generic detect changes method
Description
===========

This patch adds a new method:
`B detectChanges(String fromCommitish, String toCommitish, String branchName) throws ChangeDetectorException`

The implementation allows to pass anything that can be converted into a
commit to be passed as `from` and `to`. At the moment _tag name_, _branche name_
and _commit sha_ values are supported. This comes with a slight
performance and API call overhead because the implementation needs to
check multiple endpoints if the provided value can be converted into a
tag or branch.

Description
===========

* ![ADD] generic detect changes method