Fix fetch commits after github API update
Description
===========

The github Java API broke slightly after the latest update.
We used to normalize the commit dates to UTC before passing them
to the API. The github API is now handling this themselves and so
it broke the unit/integration tests as a timeframe of 2h (berlin local time)
is already enough to break most tests.
We still need the normalization to fetch the pullrequests because
we use the query API directly and because the commit dates handed down
from the API are in local time again.

Changes
=======

* ![FIX] fetch committs after github API update