Add snyk monitoring
Description
===========

This patch adds snyk monitoring to the build pipeline.
It will hook itself into the check and publish stages.

The patch also sets a dependency helper plugin
`net.wooga.cve-dependency-resolution` which applies overrides
for dependencies with know fixes for security issues.

Along with the introduction of snyk I also upgraded/removed some
depdencies. Coveralls produces a huge load of errors even with the
latest version so I decided to remove it since we want/are moving
to sonarqube (It is unknown at this time if this dependency is actually better
or not). I updated SpockCore and groovy-all which lead to a weird error with the
`format` method on `Date` object coming from groovy. I replaced these with
`SimpleDateFormat`

Changes
=======

* ![ADD] `snyk` monitoring
* ![ADD] `net.wooga.snyk-wdk-java` snyk convention plugin
* ![ADD] `net.wogoa.cve-dependency-resolution` plugin
* ![REMOVE] coveralls plugin
* ![UPDATE] `org.codehaus.groovy:groovy-all` to version `2.5.14`
* ![UPDATE] `org.spockframework:spock-core` to version `1.3-groovy-2.5`
* ![UPDATE] `org.ajoberstar.grgit:grgit-core` to version `4.1.1`
* ![UPDATE] `org.kohsuke:github-api` to version `1.302`